+++ examples/sha2_crates_io/kernels/src/lib.rs
@@ -5,7 +5,7 @@ use sha2::{Digest, Sha256, Sha512};
 #[kernel]
 #[allow(improper_ctypes_definitions, clippy::missing_safety_doc)]
 pub unsafe fn sha256_oneshot(input: &[u8], output: *mut [u8; 32]) {
-    let idx = thread::index_1d() as usize;
+    let idx = thread::index_1d();
 
     if idx == 0 {
         let hash = Sha256::digest(input);
@@ -21,7 +21,7 @@ pub unsafe fn sha256_oneshot(input: &[u8], output: *mut [u8; 32]) {
 #[kernel]
 #[allow(improper_ctypes_definitions, clippy::missing_safety_doc)]
 pub unsafe fn sha256_incremental(input1: &[u8], input2: &[u8], output: *mut [u8; 32]) {
-    let idx = thread::index_1d() as usize;
+    let idx = thread::index_1d();
 
     if idx == 0 {
         let mut hasher = Sha256::new();
@@ -40,7 +40,7 @@ pub unsafe fn sha256_incremental(input1: &[u8], input2: &[u8], output: *mut [u8;
 #[kernel]
 #[allow(improper_ctypes_definitions, clippy::missing_safety_doc)]
 pub unsafe fn sha512_oneshot(input: &[u8], output: *mut [u8; 64]) {
-    let idx = thread::index_1d() as usize;
+    let idx = thread::index_1d();
 
     if idx == 0 {
         let hash = Sha512::digest(input);
@@ -56,7 +56,7 @@ pub unsafe fn sha512_oneshot(input: &[u8], output: *mut [u8; 64]) {
 #[kernel]
 #[allow(improper_ctypes_definitions, clippy::missing_safety_doc)]
 pub unsafe fn sha512_incremental(input: &[u8], output: *mut [u8; 64]) {
-    let idx = thread::index_1d() as usize;
+    let idx = thread::index_1d();
 
     if idx == 0 {
         let mut hasher = Sha512::new();
